service/backup: split manifest load and deletion
Loading all manifests in parallel can be memory-intensive. This splits
the loading of manifests and the actual file deletion in the purge step.
This way by default manifests are loaded only one at a time, but a
configurable parallelism can be provided if memory is not as constrained
and they can be loaded in parallel again.

Parallelism is set per-task. For tasks created in previous versions with
no value set it will use the zero-value, which is identical to old
behavior of no limit. Newly created tasks default to loading manifests
from only 1 host at a time.

When deleting a snapshot with sctool a manifest-parallelism flag can be
supplied to override the default value of doing them sequentially.

fixes #3113