service/{backup,service}: provide ETA in progress
Using the historical data for a given task, calculate an ETA when a user
requests task progress.

For repairs this is based on # of token ranges remaining, and how
quickly they were historically repaired.

For backups this is based on bytes remaining to upload and how quickly
they historically uploaded.

Fixes #3137