service/backup: use jsoniter to stream manifest indexes
When adding streaming of manifest indexes to the purge process, we
stopped loading in the metadata for manifest content which is used in
recording progress.

This replaces the seek and decode functionality added to jsonutil with
the json-iter package. This allows us to easily decode the entire JSON
contents but still stream the indexes and retain the memory footprint
improvements.

Revert "util/jsonutil: add decoder wrapper with seeker"

This reverts commit 205788b2513cc47e869b1ead3c79f217332c0194.

fixes #3169